refactor: divided the files in domain/model into separate files.
